requirements.txt
actions==1.0.3
bunch==1.0.1
colorama==0.3.7
colorlog==2.10.0
Django==1.8.17
django-crispy-forms==1.6.1
django-el-pagination==3.0.1
django-generic-positions==0.2.2
django-hvad==1.6.0
django-libs==1.67.8
django-nocaptcha-recaptcha==0.0.19
django-user-media==1.2.3
easy-thumbnails==2.3
future==0.16.0
oauthlib==2.0.1
Pillow==3.4.2
psycopg2==2.6.1
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.3.3
requests==2.12.4
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
social-auth-core==0.2.1
unicode-slugify==0.1.3
Unidecode==0.4.19
